🐛 修复移动端在特定浏览器上上下滑动时出现白屏的问题
